Add files via upload
Implementation of a real-time control strategy that combines self-triggered control with Control Lyapunov Functions
Implementation of the controller that overcomes the main limitations of traditional approaches based on periodic controllers, i.e. unnecessary controller updates and potential violations of the safety constraints.
Central to this approach is the notion of a safe period, which enforces a safety guarantee for implementing ZOH control.